remove iodata
1. Using utils functions to perform io
2. fix all tests problems with new path
3. adjust tests files to each module/test/data
4. add __init__ to make resource folder importable

+++ saddle/utils.py
@@ -48,7 +48,7 @@ class Utils():
             size = int(f.readline())
             title = f.readline().strip()
             numbers = np.zeros(size)
-            coordinates = np.zeros((size, 3))
+            coordinates = np.zeros((size, 3), dtype=float)
             for i in range(size):
                 contents = f.readline().split()
                 numbers[i] = periodic[contents[0]].number
